+++ android/src/oldarch/java/com/ft/sdk/reactnative/FTMobileModule.java
@@ -1,20 +1,22 @@
 package com.ft.sdk.reactnative;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
+import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 
 public class FTMobileModule extends ReactContextBaseJavaModule {
+  private final FTMobileImpl impl;
 
   public FTMobileModule(ReactApplicationContext reactContext) {
     super(reactContext);
+    impl = new FTMobileImpl(reactContext);
   }
-
-  private final FTMobileImpl impl = new FTMobileImpl();
 
   @ReactMethod
   public void sdkConfig(ReadableMap context, Promise promise) {
@@ -59,6 +61,26 @@ public class FTMobileModule extends ReactContextBaseJavaModule {
   @ReactMethod
   public void clearAllData(Promise promise) {
     impl.clearAllData(promise);
+  }
+
+  @ReactMethod
+  public void updateRemoteConfig(Promise promise) {
+    impl.updateRemoteConfig(promise);
+  }
+
+  @ReactMethod
+  public void updateRemoteConfigWithMiniUpdateInterval(int interval, @Nullable ReadableArray rules, Promise promise) {
+    impl.updateRemoteConfigWithMiniUpdateInterval(interval, rules, promise);
+  }
+
+  @ReactMethod
+  public void addListener(String eventName) {
+    impl.addListener(eventName);
+  }
+
+  @ReactMethod
+  public void removeListeners(double count) {
+    impl.removeListeners(count);
   }
 
   @NonNull