+++ example/index.tsx
@@ -8,6 +8,7 @@ import {
   DetectFrequency,
   DeviceMetricsMonitorType,
   EnvType,
+  IOSCrashMonitoringType,
   ErrorMonitorType,
   FTLogConfig,
   FTLogStatus,
@@ -58,10 +59,33 @@ async function reactNativeInitSDK() {
     enableLimitWithDbSize:true,
     dbCacheLimit:50*1024*1024,
     dbDiscardStrategy:FTDBCacheDiscard.discard,
+    remoteConfiguration:true,
     // envType:EnvType.prod,
     globalContext: { 'sdk_example': 'example1' },
+    remoteConfigOverrideRules: [
+      {
+        id: 'auto_rule1',
+        enabled: true,
+        match: {
+          customKeys: {
+            env: 'test'
+          }
+        },
+        override: {
+          rumSampleRate: 1.0,
+          traceSampleRate: 1.0,
+          logSampleRate: 1.0,
+        }
+      }
+    ]
   };
+  console.log('remote config override rules configured', config.remoteConfigOverrideRules);
   await FTMobileReactNative.sdkConfig(config);
+  FTMobileReactNative.addRemoteConfigListener(
+    (result) => {
+      console.log('addRemoteConfigListener auto remote config callback', result);
+    }
+  );
 
   // log settings
   let logConfig: FTLogConfig = {
@@ -81,6 +105,7 @@ async function reactNativeInitSDK() {
     traceType: TraceType.ddTrace,
   };
   await FTReactNativeTrace.setConfig(traceConfig);
+  FTMobileReactNative.appendBridgeContext({"wgt_id":"widget_id"});
 
   // rum settings
   let rumConfig: FTRUMConfig = {
@@ -91,6 +116,8 @@ async function reactNativeInitSDK() {
     enableNativeUserAction: true,
     enableNativeUserView: false,
     sampleRate:1,
+    enableTraceWebView: true,
+    iosCrashMonitoringType: IOSCrashMonitoringType.all,
     enableNativeUserResource: true,
     enableResourceHostIP:true,
     enableTrackNativeAppANR:true,