+++ ios/FTReactNativeSessionReplay.mm
@@ -1,6 +1,5 @@
 #import "FTReactNativeSessionReplay.h"
 #import <FTMobileSDK/FTRumSessionReplay.h>
-#import <FTMobileSDK/FTSessionReplayConfig.h>
 #import <FTMobileSDK/FTSessionReplayConfig+Private.h>
 #import <React/RCTConvert.h>
 #import "FTRCTTextViewRecorder.h"
@@ -29,18 +28,34 @@ RCT_REMAP_METHOD(sessionReplayConfig,
   if([context.allKeys containsObject:@"sampleRate"]){
     config.sampleRate = [RCTConvert double:context[@"sampleRate"]]*100;
   }
+  if([context.allKeys containsObject:@"sessionReplayOnErrorSampleRate"]){
+    config.sessionReplayOnErrorSampleRate = [RCTConvert double:context[@"sessionReplayOnErrorSampleRate"]]*100;
+  }
   if([context.allKeys containsObject:@"privacy"]){
     int privacy = [context[@"privacy"] intValue];
-    switch (privacy){
-      case 0:
-        config.privacy = FTSRPrivacyMask;
-        break;
-      case 1:
-        config.privacy = FTSRPrivacyAllow;
-        break;
-      case 2:
-        config.privacy = FTSRPrivacyMaskUserInput;
-        break;
+    config.privacy = (FTSRPrivacy)privacy;
+  }
+  
+  // Handle fine-grained privacy settings (overrides deprecated privacy setting if provided)
+  if([context.allKeys containsObject:@"touchPrivacy"]){
+    int touchPrivacy = [context[@"touchPrivacy"] intValue];
+    config.touchPrivacy = (FTTouchPrivacyLevel)touchPrivacy;
+  }
+  
+  if([context.allKeys containsObject:@"textAndInputPrivacy"]){
+    int textAndInputPrivacy = [context[@"textAndInputPrivacy"] intValue];
+    config.textAndInputPrivacy = (FTTextAndInputPrivacyLevel)textAndInputPrivacy;
+  }
+  
+  if([context.allKeys containsObject:@"imagePrivacy"]){
+    int imagePrivacy = [context[@"imagePrivacy"] intValue];
+    config.imagePrivacy = (FTImagePrivacyLevel)imagePrivacy;
+  }
+  
+  if([context.allKeys containsObject:@"enableLinkRUMKeys"]){
+    NSArray *rumKeys = context[@"enableLinkRUMKeys"];
+    if([rumKeys isKindOfClass:[NSArray class]]){
+      config.enableLinkRUMKeys = rumKeys;
     }
   }
   FTRCTTextViewRecorder *recorder = [[FTRCTTextViewRecorder alloc]initWithUIManager:_bridge.uiManager];